Abstract setuptools asset compiling
