+++ src/sentry/__init__.py
@@ -10,6 +10,8 @@ from __future__ import absolute_import
 import os
 import os.path
 
+from subprocess import check_output
+
 try:
     VERSION = __import__('pkg_resources') \
         .get_distribution('sentry').version
@@ -18,14 +20,15 @@ except Exception as e:
 
 
 def _get_git_revision(path):
-    revision_file = os.path.join(path, 'refs', 'heads', 'master')
-    if not os.path.exists(revision_file):
+    if not os.path.exists(os.path.join(path, '.git')):
         return None
-    fh = open(revision_file, 'r')
     try:
-        return fh.read().strip()
-    finally:
-        fh.close()
+        revision = check_output(['git', 'rev-parse', 'HEAD'],
+                                cwd=path, env=os.environ)
+    except Exception:
+        # binary didn't exist, wasn't on path, etc
+        return None
+    return revision.strip()
 
 
 def get_revision():
@@ -37,7 +40,7 @@ def get_revision():
         return os.environ['SENTRY_BUILD']
     package_dir = os.path.dirname(__file__)
     checkout_dir = os.path.normpath(os.path.join(package_dir, os.pardir, os.pardir))
-    path = os.path.join(checkout_dir, '.git')
+    path = os.path.join(checkout_dir)
     if os.path.exists(path):
         return _get_git_revision(path)
     return None
