+++ src/sentry/utils/distutils/__init__.py
@@ -1,0 +1,4 @@
+from __future__ import absolute_import
+
+from .commands.build_integration_docs import BuildIntegrationDocsCommand  # NOQA
+from .commands.build_assets import BuildAssetsCommand  # NOQA
